+++ cpp/src/arrow/util/bpacking_simd128_generated.h
@@ -36,13 +36,7 @@ using ::arrow::util::SafeLoad;
 template <DispatchLevel level>
 struct UnpackBits128 {
 
-#ifdef ARROW_HAVE_NEON
-using simd_arch = xsimd::neon64;
-#else
-using simd_arch = xsimd::sse4_2;
-#endif
-
-using simd_batch = xsimd::batch<uint32_t, simd_arch>;
+using simd_batch = xsimd::make_sized_batch_t<uint32_t, 4>;
 
 inline static const uint32_t* unpack0_32(const uint32_t* in, uint32_t* out) {
   memset(out, 0x0, 32 * sizeof(*out));