+++ r/tests/testthat/test-dplyr-funcs-datetime.R
@@ -711,3 +711,41 @@ test_that("am/pm mirror lubridate", {
   )
 
 })
+
+test_that("extract tz", {
+  df <- tibble(
+    posixct_date = as.POSIXct(c("2022-02-07", "2022-02-10"), tz = "Pacific/Marquesas"),
+  )
+
+  compare_dplyr_binding(
+    .input %>%
+      mutate(timezone_posixct_date = tz(posixct_date)) %>%
+      collect(),
+    df
+  )
+
+  # test a few types directly from R objects
+  expect_error(
+    call_binding("tz", "2020-10-01"),
+    "timezone extraction for objects of class `string` not supported in Arrow"
+  )
+  expect_error(
+    call_binding("tz", as.Date("2020-10-01")),
+    "timezone extraction for objects of class `date32[day]` not supported in Arrow",
+    fixed = TRUE
+  )
+  expect_error(
+    call_binding("tz", 1L),
+    "timezone extraction for objects of class `int32` not supported in Arrow"
+  )
+   expect_error(
+    call_binding("tz", 1.1),
+    "timezone extraction for objects of class `double` not supported in Arrow"
+  )
+
+  # Test one expression
+  expect_error(
+    call_binding("tz", Expression$scalar("2020-10-01")),
+    "timezone extraction for objects of class `string` not supported in Arrow"
+  )
+})